requirements.txt
streamlit==1.35.0
PyMuPDF==1.24.1
spacy==3.7.4
sentence-transformers==2.6.1
faiss-cpu==1.7.4
torch==2.3.0
numpy==1.26.4
